Fixed FastMath cpp realization

+++ FastExponentiation/FastMathCpp/FastMath.h
@@ -3,9 +3,9 @@
 
 static class FastMath {
 public:
-	static double BinaryPower(double b, long e);
+	static double BinaryPower(double b, long long e);
 
-	static const long doubleApproximator = 4606853616395542500L;
+	static const long long doubleApproximator = 4606853616395542500L;
 	static double FastApproximatePower(double b, double e);
 
 	static double FastPower(double b, double e);
